Moved neptune.types.mode to neptune.objects
