+++ src/neptune/objects/model_version.py
@@ -47,11 +47,11 @@ from neptune.internal.parameters import (
 from neptune.internal.state import ContainerState
 from neptune.internal.utils import verify_type
 from neptune.internal.utils.ping_background_job import PingBackgroundJob
+from neptune.objects.mode import Mode
 from neptune.objects.neptune_object import (
     NeptuneObject,
     NeptuneObjectCallback,
 )
-from neptune.types.mode import Mode
 from neptune.types.model_version_stage import ModelVersionStage
 
 if TYPE_CHECKING:
